cashflow/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
PyJWT
psycopg2
kafka-python
